+++ src/core/solution.cpp
@@ -48,7 +48,7 @@ void Solution::set_composition(const Eigen::ArrayXd& composition_vector) {
     throw std::runtime_error("Cannot set molar fractions: solution model not set!");
   }
   // Throw error if length not correct
-  if (composition_vector.size() != get_n_endmembers()) {
+  if (static_cast<int>(composition_vector.size()) != get_n_endmembers()) {
     throw std::runtime_error(
       "Composition vector length (" + std::to_string(composition_vector.size()) +
       ") does not match number of endmembers (" + std::to_string(get_n_endmembers()) + ").");
